[rest] casos de erros, validacao tratados

+++ libraryapi/src/main/java/com/world/haymon/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package com.world.haymon.libraryapi.service;
 
+import com.world.haymon.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.world.haymon.libraryapi.model.Autor;
 import com.world.haymon.libraryapi.repository.AutorRepository;
+import com.world.haymon.libraryapi.repository.LivroRepository;
+import com.world.haymon.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,19 +15,25 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository) {
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
+        validator.validar(autor);
         return repository.save(autor);
     }
 
     public void atualizar(Autor autor) {
         if (autor.getId() == null) throw new IllegalArgumentException(
-                "Não é possível atualizar Autor inexistente!"
+                "Não é possível atualizar Autor que não existe!"
         );
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -33,6 +42,11 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException(
+                    "Não é possível excluir um Autor que possui Livros!"
+            );
+        }
         repository.delete(autor);
     }
 
@@ -41,5 +55,9 @@ public class AutorService {
         if (nome != null) return repository.findByNome(nome);
         if (nacionalidade != null) return repository.findByNacionalidade(nacionalidade);
         return repository.findAll();
+    }
+
+    private boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
